+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * The **modeling-rules test** command will now validate that the modeling rules schema mappings are aligned with the test-data mappings.
 * Added the *--xsiam* flag to the **init** command in order to create XSIAM content.
 * Removed the format execution step from the `contribution_converter` since it can be executed separately during the contribution process.
+* Added a new validation (`GR108`) to **validate** command that assures hidden packs do not have mandatory dependant packs.
 * Added a validation that check that non ignorable errors are not used.
 
 ## 1.18.1